Feature/upgrade from hb 0.32

+++ hummingbot/strategy/liquidity_mirroring/liquidity_mirroring_config_map.py
@@ -86,56 +86,56 @@ liquidity_mirroring_config_map = {
         prompt="Enter your desired order price markup applied after fees (0.01 is 1%) >>> ",
         default=0.001,
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "max_exposure_base": ConfigVar(
         key="max_exposure_base",
         prompt="Enter the max desired exposure for the base asset >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "max_exposure_quote": ConfigVar(
         key="max_exposure_quote",
         prompt="Enter the max desired exposure for the quote asset >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "max_offsetting_exposure": ConfigVar(
         key="max_offsetting_exposure",
         prompt="Enter the maximum allowable outstanding amount to offset >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "max_offset_loss": ConfigVar(
         key="max_offset_loss",
         prompt="Enter your maximum tolerated one-time percentage offset loss (0.01 for 1%) >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "max_total_offset_loss": ConfigVar(
         key="max_total_offset_loss",
         prompt="Enter your maximum tolerated total offset loss (in quote currency) >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "min_primary_amount": ConfigVar(
         key="min_primary_amount",
         prompt="Enter the min amount of base currency per trade on the primary exchange >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "min_mirroring_amount": ConfigVar(
         key="min_mirroring_amount",
         prompt="Enter the min amount of base currency per trade on the mirrored exchange >>> ",
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "order_replacement_threshold": ConfigVar(
         key="order_replacement_threshold",
         prompt=("Enter the min percentage change in best bid/ask on taker exchange to trigger " 
           "an order replacment for best bid/ask on the maker exchange (1% = 0.01) >>> "),
         prompt_on_new=True,
-        type_str="float"
+        type_str="decimal"
     ),
     "bid_amount_ratio_type": ConfigVar(
         key="bid_amount_ratio_type",
@@ -158,7 +158,7 @@ liquidity_mirroring_config_map = {
         prompt="Enter the period of time between slack status updates (in hours) >>> ",
         prompt_on_new=True,
         default=1.0,
-        type_str="float"
+        type_str="decimal"
     ),
     "ask_amount_ratios": ConfigVar(
         key="ask_amount_ratios",
